+++ src/Iced/Widget/Checkbox.hs
@@ -28,7 +28,7 @@ import Iced.Attribute.Size
 import Iced.Attribute.Spacing
 import Iced.Attribute.Status
 import Iced.Attribute.Style
-import Iced.Attribute.TextFFI
+import Iced.Attribute.Text
 import Iced.ColorFFI
 import Iced.Element
 import Iced.Theme