+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## Unreleased
 
 * Add automatic formatting
+* Added canvas frame methods: fillText, pushTransform, popTransform, rotate, scale